# lib/api-zod/src/generated/types/steldexAddLiquidityInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexAddLiquidityInput {
  walletAddress: string;
  poolContract: string;
  token0Contract: string;
  token1Contract: string;
  tickLower: number;
  tickUpper: number;
  amount0Desired: string;
  amount1Desired: string;
  /** @nullable */
  stepId?: string | null;
}