# lib/api-zod/src/generated/types/steldexClaimInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexClaimInput {
  walletAddress: string;
  poolContract: string;
  tickLower: number;
  tickUpper: number;
}