# lib/api-zod/src/generated/types/steldexFarmPool.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { SteldexFarmPoolFarm } from './steldexFarmPoolFarm';

export interface SteldexFarmPool {
  /** @nullable */
  pair?: string | null;
  /** @nullable */
  poolContract?: string | null;
  /** @nullable */
  token0Symbol?: string | null;
  /** @nullable */
  token1Symbol?: string | null;
  /** @nullable */
  tvlUsd?: number | null;
  /** @nullable */
  lpLiquidity?: string | null;
  /** @nullable */
  availableToStake?: string | null;
  /** @nullable */
  stakedLiquidity?: string | null;
  /** @nullable */
  tickLower?: number | null;
  /** @nullable */
  tickUpper?: number | null;
  farm?: SteldexFarmPoolFarm;
  [key: string]: unknown;
 }