# lib/api-zod/src/generated/types/steldexLimitOrderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { SteldexLimitOrderInputOrderType } from './steldexLimitOrderInputOrderType';

export interface SteldexLimitOrderInput {
  walletAddress: string;
  fromContract: string;
  toContract: string;
  amount: string;
  limitPrice: string;
  orderType: SteldexLimitOrderInputOrderType;
  /** @nullable */
  expiryHours?: number | null;
  /** @nullable */
  stepId?: string | null;
}