# lib/api-zod/src/generated/types/steldexOrder.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexOrder {
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  pair?: string | null;
  /** @nullable */
  status?: string | null;
  [key: string]: unknown;
 }