# lib/api-zod/src/generated/types/steldexStakeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexStakeInput {
  walletAddress: string;
  poolContract: string;
  tickLower: number;
  tickUpper: number;
  /** @nullable */
  stakeMax?: boolean | null;
  /** @nullable */
  liquidity?: string | null;
  lockWeeks: number;
  /** @nullable */
  autoCompound?: boolean | null;
  /** @nullable */
  stepId?: string | null;
}