# lib/api-zod/src/generated/types/steldexSwapQuoteInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexSwapQuoteInput {
  fromTokenContract: string;
  toTokenContract: string;
  amountIn: string;
  /** @nullable */
  slippageBps?: number | null;
}