# lib/api-zod/src/generated/types/steldexSwapQuoteResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexSwapQuoteResult {
  /** @nullable */
  amountOut?: string | null;
  /** @nullable */
  minAmountOut?: string | null;
  /** @nullable */
  priceImpact?: number | null;
  [key: string]: unknown;
 }