# lib/api-zod/src/generated/types/steldexTxResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { SteldexTxStep } from './steldexTxStep';

export interface SteldexTxResult {
  /** @nullable */
  xdr?: string | null;
  /** @nullable */
  steps?: SteldexTxStep[] | null;
  /** @nullable */
  sequential?: boolean | null;
  /** @nullable */
  resting?: boolean | null;
  [key: string]: unknown;
 }