# lib/api-zod/src/generated/types/steldexTxStep.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexTxStep {
  id?: string;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  xdr?: string | null;
  [key: string]: unknown;
 }