# lib/api-zod/src/generated/types/steldexUnstakeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SteldexUnstakeInput {
  walletAddress: string;
  poolContract: string;
  tickLower: number;
  tickUpper: number;
  /** @nullable */
  unstakeMax?: boolean | null;
  /** @nullable */
  liquidity?: string | null;
  /** @nullable */
  stepId?: string | null;
}